server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==15.0.1
soundfile==0.13.1

# Keep <2 due to NeMo ASR headaches with NumPy 2.x
numpy==1.26.4
scipy==1.14.1

# Torch stays in Dockerfile; ensure versions match:
# e.g., torch==2.3.1+cu121  ↔  torchaudio==2.3.1
torchaudio==2.3.1
silero-tts

nemo_toolkit[asr]==2.4.0
pytest==8.4.1
pytest-asyncio==1.1.0
